+++ src/util/subgraph.c
@@ -84,7 +84,7 @@ void WriteSubgraph(Subgraph *g, char *file){
 
     /*writing position(id), label and features*/
     for (i = 0; i < g->nnodes; i++){
-      fwrite(&g->node[i].position, sizeof(long int), 1, fp);
+      fwrite(&g->node[i].position, sizeof(int), 1, fp);
       fwrite(&g->node[i].truelabel, sizeof(int), 1, fp);
       for (j = 0; j < g->nfeats; j++)
 	fwrite(&g->node[i].feat[j], sizeof(float), 1, fp);
@@ -117,7 +117,7 @@ Subgraph *ReadSubgraph(char *file){
   /*reading features*/
   for (i = 0; i < g->nnodes; i++){
     g->node[i].feat = AllocFloatArray(g->nfeats);
-    if (fread(&g->node[i].position, sizeof(long int), 1, fp) != 1) 
+    if (fread(&g->node[i].position, sizeof(int), 1, fp) != 1) 
       Error("Could not read node position","ReadSubGraph");      
     if(fread(&g->node[i].truelabel, sizeof(int), 1, fp) != 1) 
       Error("Could not read node true label","ReadSubGraph");
